# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alert incidents
      parameters:
        - { name: state, in: query, schema: { $ref: "#/components/schemas/AlertState" } }
        - { name: severity, in: query, schema: { $ref: "#/components/schemas/AlertSeverity" } }
        - { name: type, in: query, schema: { $ref: "#/components/schemas/AlertType" } }
        - { name: device_id, in: query, schema: { type: string, format: uuid } }
        - { name: organization, in: query, schema: { type: string } }
        - { name: active, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        "200":
          description: Paginated alerts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertList" }
        "400": { description: Invalid filters }
  /v1/alerts/{alert_id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get alert detail and timeline
      parameters:
        - { name: alert_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Alert detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertDetail" }
        "404": { description: Alert not found }
  /v1/alerts/{alert_id}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: Acknowledge an open alert
      security: [{ adminBearer: [] }]
      parameters:
        - { name: alert_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Alert" }
        "401": { description: Administrative authorization required }
        "409": { description: Resolved alert cannot be acknowledged }
...
  /v1/admin/enrollment-tokens:
    get:
      operationId: listEnrollmentTokens
      tags: [administration]
      security: [{ adminBearer: [] }]
      responses:
        "200":
          description: Enrollment-token metadata without raw secrets
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/EnrollmentToken" } }
    post:
      operationId: createEnrollmentToken
      tags: [administration]
      security: [{ adminBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEnrollmentTokenInput" }
      responses:
        "201":
          description: Token created; raw token is returned only here
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedEnrollmentToken" }
  /v1/admin/enrollment-tokens/{token_id}/revoke:
    post:
      operationId: revokeEnrollmentToken
      tags: [administration]
      security: [{ adminBearer: [] }]
      parameters:
        - name: token_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Token revoked }
...
components:
...
  securitySchemes:
    agentBearer:
      type: http
      scheme: bearer
    adminBearer:
      type: http
      scheme: bearer
    # Interactive console authentication. Deliberately separate from adminBearer,
    # which is a platform-level machine credential and is never issued to a tenant.
    sessionCookie:
      type: apiKey
      in: cookie
      name: nexora_session
  schemas:
...
    AlertType:
      type: string
      enum: [DEVICE_OFFLINE, CPU_HIGH, MEMORY_HIGH, DISK_HIGH]
    AlertSeverity:
      type: string
      enum: [warning, critical]
    AlertState:
      type: string
      enum: [OPEN, ACKNOWLEDGED, RESOLVED]
    Alert:
      type: object
      required: [id, organization, device_id, device, type, severity, state, resource, title, summary, opened_at, last_triggered_at, acknowledged_at, resolved_at, acknowledged_by, trigger_value, threshold_value, occurrence_count, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        organization: { type: string }
        device_id: { type: string, format: uuid }
        device:
          type: object
          required: [id, hostname]
          properties:
            id: { type: string, format: uuid }
            hostname: { type: string }
            last_seen_at: { type: [string, "null"], format: date-time }
        type: { $ref: "#/components/schemas/AlertType" }
        severity: { $ref: "#/components/schemas/AlertSeverity" }
        state: { $ref: "#/components/schemas/AlertState" }
        resource: { type: [string, "null"] }
        title: { type: string }
        summary: { type: string }
        opened_at: { type: string, format: date-time }
        last_triggered_at: { type: string, format: date-time }
        acknowledged_at: { type: [string, "null"], format: date-time }
        resolved_at: { type: [string, "null"], format: date-time }
        acknowledged_by: { type: [string, "null"] }
        trigger_value: { type: [number, "null"] }
        threshold_value: { type: [number, "null"] }
        occurrence_count: { type: integer, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AlertEvent:
      type: object
      required: [id, event_type, previous_state, new_state, previous_severity, new_severity, actor, timestamp, metadata]
      properties:
        id: { type: string, format: uuid }
        event_type: { type: string, enum: [CREATED, SEVERITY_CHANGED, ACKNOWLEDGED, RESOLVED] }
        previous_state: { oneOf: [{ $ref: "#/components/schemas/AlertState" }, { type: "null" }] }
        new_state: { oneOf: [{ $ref: "#/components/schemas/AlertState" }, { type: "null" }] }
        previous_severity: { oneOf: [{ $ref: "#/components/schemas/AlertSeverity" }, { type: "null" }] }
        new_severity: { oneOf: [{ $ref: "#/components/schemas/AlertSeverity" }, { type: "null" }] }
        actor: { type: [string, "null"] }
        timestamp: { type: string, format: date-time }
        metadata: { type: [object, "null"], additionalProperties: true }
    AlertDetail:
      allOf:
        - { $ref: "#/components/schemas/Alert" }
        - type: object
          required: [events]
          properties:
            events: { type: array, items: { $ref: "#/components/schemas/AlertEvent" } }
    AlertList:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Alert" } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
    EnrollmentToken:
      type: object
      required: [id, name, organization_id, expires_at, max_uses, current_uses, created_at, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        organization_id: { type: string, format: uuid }
        organization_name: { type: [string, "null"] }
        site_id: { type: [string, "null"], format: uuid }
        site_name: { type: [string, "null"] }
        expires_at: { type: string, format: date-time }
        max_uses: { type: integer }
        current_uses: { type: integer }
        created_at: { type: string, format: date-time }
        revoked_at: { type: [string, "null"], format: date-time }
        active: { type: boolean }
    CreateEnrollmentTokenInput:
      type: object
      # organization_id is mandatory: every enrollment token is tenant-scoped and
      # the server assigns the enrolling device to this organization.
      required: [name, organization_id, expires_at, max_uses]
      properties:
        name: { type: string, minLength: 1 }
        organization_id: { type: string, format: uuid }
        site_id: { type: [string, "null"], format: uuid }
        expires_at: { type: string, format: date-time }
        max_uses: { type: integer, minimum: 1, maximum: 10000 }
    CreatedEnrollmentToken:
      allOf:
        - $ref: "#/components/schemas/EnrollmentToken"
        - type: object
          required: [token]
          properties:
            token: { type: string }